Time: 4 ms (38.72%) | Memory: 8.4 MB (5.16%) - LeetSync

+++ 1833-find-the-highest-altitude/find-the-highest-altitude.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int largestAltitude(vector<int>& gain) {
-        int a=0,max=0;
+        short int a=0,max=0;
         for(auto i:gain)
         {
             a+=i;
